picking the right validation method from the schema based on the currently rendered input fields
